+++ docs/source/changelog.rst
@@ -2,10 +2,25 @@
 Changelog
 =========
 
-3.0
+3.0 (TBD)
 ------------------------
 
 * Rewrite of :mod:`pyairtable.formulas` module.
+  - `PR #329 <https://github.com/gtalarico/pyairtable/pull/329>`_.
+
+2.3.2 (2024-03-18)
+------------------------
+
+* Fixed a bug affecting :func:`pyairtable.metadata.get_table_schema`.
+  - `PR #349 <https://github.com/gtalarico/pyairtable/pull/349>`_.
+
+2.3.1 (2024-03-14)
+------------------------
+
+* Fixed a bug affecting how timezones are parsed by :class:`~pyairtable.orm.fields.DatetimeField`.
+  - `PR #342 <https://github.com/gtalarico/pyairtable/pull/342>`_.
+* Fixed a bug affecting :meth:`~pyairtable.Base.create_table`.
+  - `PR #345 <https://github.com/gtalarico/pyairtable/pull/345>`_.
 
 2.3.0 (2024-02-25)
 ------------------------